% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitAutoReg-boot-ci.R
\name{BootCI}
\alias{BootCI}
\title{Bootstrap Percentile Confidence Intervals}
\usage{
BootCI(x, alpha = 0.05)
}
\arguments{
\item{x}{Numeric matrix.
Output of
\code{\link[=PBootVAROLS]{PBootVAROLS()}},
\code{\link[=PBootVARLasso]{PBootVARLasso()}},
\code{\link[=RBootVAROLS]{RBootVAROLS()}}, or
\code{\link[=RBootVARLasso]{RBootVARLasso()}}.}

\item{alpha}{Numeric.
Significance level.}
}
\value{
A list with two elements, namely \code{ll} for the lower limit
and \code{ul} for the upper limit.
}
\description{
Bootstrap Percentile Confidence Intervals
}
\examples{
set.seed(42)
# Parametric bootstrap
system.time(
  pb <- PBootVAROLS(
    data = dat_p2,
    p = 2,
    B = 10,
    burn_in = 20
  )
)
pb$est
BootCI(pb)
system.time(
  pb <- PBootVARLasso(
    data = dat_p2,
    p = 2,
    B = 10,
    burn_in = 20,
    n_lambdas = 100,
    crit = "ebic",
    max_iter = 1000,
    tol = 1e-5
  )
)
pb$est
BootCI(pb)

# Residual bootstrap
system.time(
  rb <- RBootVAROLS(
    data = dat_p2,
    p = 2,
    B = 10
  )
)
rb$est
BootCI(rb)
system.time(
  rb <- RBootVARLasso(
    data = dat_p2,
    p = 2,
    B = 10,
    n_lambdas = 100,
    crit = "ebic",
    max_iter = 1000,
    tol = 1e-5
  )
)
rb$est
BootCI(rb)

}
\seealso{
Other Simulation of Autoregressive Data Functions: 
\code{\link{BootSE}()},
\code{\link{SelectVARLasso}()},
\code{\link{SimVAR}()},
\code{\link{YX}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation of Autoregressive Data Functions}
\keyword{fitAutoReg}
\keyword{pb}
\keyword{rb}
