% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitAutoReg-boot-se.R
\name{BootSE}
\alias{BootSE}
\title{Bootstrap Standard Errors}
\usage{
BootSE(x)
}
\arguments{
\item{x}{Numeric matrix.
Output of
\code{\link[=PBootVAROLS]{PBootVAROLS()}},
\code{\link[=PBootVARLasso]{PBootVARLasso()}},
\code{\link[=RBootVAROLS]{RBootVAROLS()}}, or
\code{\link[=RBootVARLasso]{RBootVARLasso()}}.}
}
\value{
A matrix of standard errors.
}
\description{
Bootstrap Standard Errors
}
\examples{
set.seed(42)
# Parametric bootstrap
system.time(
  pb <- PBootVAROLS(
    data = dat_p2,
    p = 2,
    B = 10,
    burn_in = 20
  )
)
pb$est
BootSE(pb)
system.time(
  pb <- PBootVARLasso(
    data = dat_p2,
    p = 2,
    B = 10,
    burn_in = 20,
    n_lambdas = 100,
    crit = "ebic",
    max_iter = 1000,
    tol = 1e-5
  )
)
pb$est
BootSE(pb)

# Residual bootstrap
system.time(
  rb <- RBootVAROLS(
    data = dat_p2,
    p = 2,
    B = 10
  )
)
rb$est
BootSE(rb)
system.time(
  rb <- RBootVARLasso(
    data = dat_p2,
    p = 2,
    B = 10,
    n_lambdas = 100,
    crit = "ebic",
    max_iter = 1000,
    tol = 1e-5
  )
)
rb$est
BootSE(rb)

}
\seealso{
Other Simulation of Autoregressive Data Functions: 
\code{\link{BootCI}()},
\code{\link{SelectVARLasso}()},
\code{\link{SimVAR}()},
\code{\link{YX}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation of Autoregressive Data Functions}
\keyword{fitAutoReg}
\keyword{pb}
\keyword{rb}
