% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitAutoReg-methods.R
\name{coef.dynr_model}
\alias{coef.dynr_model}
\title{Auto and Cross Regression Coeffcients}
\usage{
\method{coef}{dynr_model}(object, dynr_cook, ...)
}
\arguments{
\item{object}{Object of class \code{dynr_model}.}

\item{dynr_cook}{Ouput of \code{\link[=FitVARDynr]{FitVARDynr()}}.}

\item{...}{additional arguments.}
}
\value{
A list with the following elements:
\itemize{
\item constant Numeric vector.
The constant term vector of length \code{k},
where \code{k} is the number of variables.
\item coef Numeric matrix.
Coefficient matrix with dimensions \code{k} by \code{(k * p)}.
Each \code{k} by \code{k} block corresponds to the coefficient matrix
for a particular lag.
\item cov Numeric matrix.
The \code{k} by \code{k} covariance matrix of the noise.
}
}
\description{
Auto and Cross Regression Coeffcients
}
\examples{
\dontrun{
dynr_model <- ModelVARP2Dynr(data = dat_demo)
dynr_cook <- FitVARP2Dynr(model = dynr_model)
coef(object = dynr_model, dynr_cook = dynr_cook)
}

}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
