% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitAutoReg-dat-demo-exo.R
\docType{data}
\name{dat_demo_exo}
\alias{dat_demo_exo}
\title{Data from the Vector Autoregressive Model with Exogenous Variables}
\format{
A matrix with 1000 rows (time points) and
k = 3 (autoregressive variables)
plus m = 3 columns (exogenous variables)
generated from the p = 2 vector autoregressive model given by
\deqn{
    Y_{1_{t}}
    =
    1 +
    0.4 Y_{1_{t - 1}} + 0.0 Y_{2_{t - 1}} + 0.0 Y_{3_{t - 1}} +
    0.1 Y_{1_{t - 2}} + 0.0 Y_{2_{t - 2}} + 0.0 Y_{3_{t - 2}} +
    0.5 X_{1} + 0.0 X_{2} + 0.0 X_{3}
   \varepsilon_{1_{t}} ,
  }
\deqn{
    Y_{2_{t}}
    =
    1 +
    0.0 Y_{1_{t - 1}} + 0.5 Y_{2_{t - 1}} + 0.0 Y_{3_{t - 1}} +
    0.0 Y_{1_{t - 2}} + 0.2 Y_{2_{t - 2}} + 0.0 Y_{3_{t - 2}} +
    0.0 X_{1} + 0.5 X_{2} + 0.0 X_{3}
    \varepsilon_{2_{t}} ,
  } and
\deqn{
    Y_{3_{t}}
    =
    1 +
    0.0 Y_{1_{t - 1}} + 0.0 Y_{2_{t - 1}} + 0.6 Y_{3_{t - 1}} +
    0.0 Y_{1_{t - 2}} + 0.0 Y_{2_{t - 2}} + 0.3 Y_{3_{t - 2}} +
    0.0 X_{1} + 0.0 X_{2} + 0.5 X_{3}
    \varepsilon_{3_{t}}
  }
which simplifies to
\deqn{
    Y_{1_{t}} = 1 + 0.4 Y_{1_{t - 1}} + 0.1 Y_{1_{t - 2}} + 0.5 X_{1}
    \varepsilon_{1_{t}} ,
  }
\deqn{
    Y_{2_{t}} = 1 + 0.5 Y_{2_{t - 1}} + 0.2 Y_{2_{t - 2}} + 0.5 X_{2}
    \varepsilon_{2_{t}} ,
  } and
\deqn{
    Y_{3_{t}} = 1 + 0.6 Y_{3_{t - 1}} + 0.3 Y_{3_{t - 2}} + 0.5 X_{3}
    \varepsilon_{3_{t}} .
  }
The covariance matrix of process noise is an identity matrix.
}
\usage{
dat_demo_exo
}
\description{
Data from the Vector Autoregressive Model with Exogenous Variables
}
\keyword{data}
\keyword{fitAutoReg}
