% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitAutoReg-dat-demo-yx.R
\docType{data}
\name{dat_demo_yx}
\alias{dat_demo_yx}
\title{Data from the Vector Autoregressive Model (Y) and Lagged Predictors (X)}
\format{
A list with elements Y and X where Y is equal to the \code{dat_demo}
data set minus p = 2 terminal rows
and \code{X} is a matrix of ones for the first column
and lagged values of \code{Y} for the rest of the columns.
}
\usage{
dat_demo_yx
}
\description{
Data from the Vector Autoregressive Model (Y) and Lagged Predictors (X)
}
\keyword{data}
\keyword{fitAutoReg}
