% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitAutoReg-dat-p1-yx.R
\docType{data}
\name{dat_p1_yx}
\alias{dat_p1_yx}
\title{Data from the Vector Autoregressive Model (Y) and Lagged Predictors (X)
(p = 1)}
\format{
A list with elements Y and X where Y is equal to the \code{dat_p1}
data set minus p = 1 terminal rows
and \code{X} is a matrix of ones for the first column
and lagged values of \code{Y} for the rest of the columns.
}
\usage{
dat_p1_yx
}
\description{
Data from the Vector Autoregressive Model (Y) and Lagged Predictors (X)
(p = 1)
}
\keyword{data}
\keyword{fitAutoReg}
