% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitAutoReg-dat-p2-exo-yx.R
\docType{data}
\name{dat_p2_exo_yx}
\alias{dat_p2_exo_yx}
\title{Data from the Vector Autoregressive Model (Y) and Lagged Predictors
and Exogenous Variables (X)  (p = 2)}
\format{
A list with elements Y and X where Y
is equal to the k = 3 autoregressive variables
of the \code{dat_p2_exo}
data set minus p = 2 terminal rows
and \code{X} is a matrix of ones for the first column,
lagged values of \code{Y}, and m = 3 exogenous variables.
}
\usage{
dat_p2_exo_yx
}
\description{
Data from the Vector Autoregressive Model (Y) and Lagged Predictors
and Exogenous Variables (X)  (p = 2)
}
\keyword{data}
\keyword{fitAutoReg}
