% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitAutoReg-package.R
\docType{package}
\name{fitAutoReg-package}
\alias{fitAutoReg-package}
\alias{_PACKAGE}
\title{fitAutoReg: Fit Autoregressive Models}
\description{
Fit autoregressive models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ijapesigan/fitAutoReg}
  \item \url{https://ijapesigan.github.io/fitAutoReg/}
  \item Report bugs at \url{https://github.com/ijapesigan/fitAutoReg/issues}
}

}
\author{
\strong{Maintainer}: Ivan Jacob Agaloos Pesigan \email{ijapesigan@gmail.com} (\href{https://orcid.org/0000-0003-4818-8420}{ORCID}) [copyright holder]

}
\keyword{internal}
