% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitAutoReg-fit-ml-var-dynr.R
\name{FitMLVARDynr}
\alias{FitMLVARDynr}
\title{Fit Vector Autoregressive (VAR) Model using dynr
on Each of the Data Matrix in a List}
\usage{
FitMLVARDynr(
  data,
  p,
  ncores = 1,
  conf_level = 0.95,
  optimization_flag = TRUE,
  hessian_flag = TRUE,
  verbose = FALSE,
  weight_flag = FALSE,
  debug_flag = FALSE,
  perturb_flag = FALSE
)
}
\arguments{
\item{data}{List.
Each element is a numeric matrix of
time series data with dimensions \code{t} by \code{k},
where \code{t} is the number of observations
and \code{k} is the number of variables.}

\item{p}{Positive integer.
Number of lags.
Only supports \code{p = 1} and \code{p = 2}.}

\item{ncores}{Positive integer.
Number of cores to use.
Not supported on Windows.}

\item{conf_level}{a  cumulative proportion indicating
the level of desired confidence intervals
for the final parameter estimates (default is .95)}

\item{optimization_flag}{a flag (TRUE/FALSE) indicating whether optimization is to be done.}

\item{hessian_flag}{a flag (TRUE/FALSE) indicating whether the Hessian matrix is to be calculated.}

\item{verbose}{a flag (TRUE/FALSE) indicating whether more detailed intermediate output during the 
estimation process should be printed}

\item{weight_flag}{a flag (TRUE/FALSE) indicating whether the negative log likelihood function should 
be weighted by the length of the time series for each individual}

\item{debug_flag}{a flag (TRUE/FALSE) indicating whether users want additional dynr output that can 
be used for diagnostic purposes}

\item{perturb_flag}{a flag (TRUE/FLASE) indicating whether to perturb the latent states during estimation. Only useful for ensemble forecasting.}
}
\value{
A list each element of which
is an object of class \code{dynrCook}.
}
\description{
This function estimates the parameters of a VAR model
using the \code{dynr} package
for each of the data matrix in a list.
}
\seealso{
Other Fitting Autoregressive Model Functions: 
\code{\link{FitVARDynr}()},
\code{\link{FitVARLassoSearch}()},
\code{\link{FitVARLasso}()},
\code{\link{FitVAROLS}()},
\code{\link{LambdaSeq}()},
\code{\link{ModelVARP1Dynr}()},
\code{\link{ModelVARP2Dynr}()},
\code{\link{OrigScale}()},
\code{\link{PBootVARLasso}()},
\code{\link{PBootVAROLS}()},
\code{\link{SearchVARLasso}()},
\code{\link{StdMat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Fitting Autoregressive Model Functions}
\keyword{fit}
\keyword{fitAutoReg}
