% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitAutoReg-fit-var-dynr.R
\name{FitVARDynr}
\alias{FitVARDynr}
\title{Fit Vector Autoregressive Model using dynr}
\usage{
FitVARDynr(
  model,
  conf_level = 0.95,
  optimization_flag = TRUE,
  hessian_flag = FALSE,
  verbose = FALSE,
  weight_flag = FALSE,
  debug_flag = FALSE,
  perturb_flag = FALSE
)
}
\arguments{
\item{model}{Ouput of \code{\link[=ModelVARP1Dynr]{ModelVARP1Dynr()}} or \code{\link[=ModelVARP2Dynr]{ModelVARP2Dynr()}}.}

\item{conf_level}{a  cumulative proportion indicating
the level of desired confidence intervals
for the final parameter estimates (default is .95)}

\item{optimization_flag}{a flag (TRUE/FALSE) indicating whether optimization is to be done.}

\item{hessian_flag}{a flag (TRUE/FALSE) indicating whether the Hessian matrix is to be calculated.}

\item{verbose}{a flag (TRUE/FALSE) indicating whether more detailed intermediate output during the 
estimation process should be printed}

\item{weight_flag}{a flag (TRUE/FALSE) indicating whether the negative log likelihood function should 
be weighted by the length of the time series for each individual}

\item{debug_flag}{a flag (TRUE/FALSE) indicating whether users want additional dynr output that can 
be used for diagnostic purposes}

\item{perturb_flag}{a flag (TRUE/FLASE) indicating whether to perturb the latent states during estimation. Only useful for ensemble forecasting.}
}
\value{
Object of class \code{dynrCook}.
}
\description{
This function estimates the parameters of a VAR model
using the \code{dynr} package.
}
\examples{
\dontrun{
FitVARDynr(model = ModelVARP1Dynr(data = dat_p1))
FitVARDynr(model = ModelVARP2Dynr(data = dat_p2))
}

}
\seealso{
Other Fitting Autoregressive Model Functions: 
\code{\link{FitMLVARDynr}()},
\code{\link{FitMLVARMplus}()},
\code{\link{FitVARLassoSearch}()},
\code{\link{FitVARLasso}()},
\code{\link{FitVARMplus}()},
\code{\link{FitVAROLS}()},
\code{\link{LambdaSeq}()},
\code{\link{ModelVARP1Dynr}()},
\code{\link{ModelVARP2Dynr}()},
\code{\link{OrigScale}()},
\code{\link{PBootVARExoLasso}()},
\code{\link{PBootVARExoOLS}()},
\code{\link{PBootVARLasso}()},
\code{\link{PBootVAROLS}()},
\code{\link{RBootVARExoLasso}()},
\code{\link{RBootVARExoOLS}()},
\code{\link{RBootVARLasso}()},
\code{\link{RBootVAROLS}()},
\code{\link{SearchVARLasso}()},
\code{\link{StdMat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Fitting Autoregressive Model Functions}
\keyword{fit}
\keyword{fitAutoReg}
