% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FitVARLassoSearch}
\alias{FitVARLassoSearch}
\title{Fit Vector Autoregressive (VAR) Model Parameters
using Lasso Regularization with Lambda Search}
\usage{
FitVARLassoSearch(YStd, XStd, lambdas, crit, max_iter, tol)
}
\arguments{
\item{YStd}{Numeric matrix.
Matrix of standardized dependent variables (Y).}

\item{XStd}{Numeric matrix.
Matrix of standardized predictors (X).
\code{XStd} should not include a vector of ones in column one.}

\item{lambdas}{Numeric vector.
Lasso hyperparameter.
The regularization strength controlling the sparsity.}

\item{crit}{Character string.
Information criteria to use.
Valid values include \code{"aic"}, \code{"bic"}, and \code{"ebic"}.}

\item{max_iter}{Integer.
The maximum number of iterations
for the coordinate descent algorithm
(e.g., \code{max_iter = 10000}).}

\item{tol}{Numeric.
Convergence tolerance.
The algorithm stops when the change in coefficients
between iterations is below this tolerance
(e.g., \code{tol = 1e-5}).}
}
\value{
Matrix of estimated autoregressive
and cross-regression coefficients.
}
\description{
Fit Vector Autoregressive (VAR) Model Parameters
using Lasso Regularization with Lambda Search
}
\examples{
YStd <- StdMat(dat_p2_yx$Y)
XStd <- StdMat(dat_p2_yx$X[, -1]) # remove the constant column
lambdas <- LambdaSeq(
  YStd = YStd,
  XStd = XStd,
  n_lambdas = 100
)
FitVARLassoSearch(
  YStd = YStd,
  XStd = XStd,
  lambdas = lambdas,
  crit = "ebic",
  max_iter = 1000,
  tol = 1e-5
)

}
\seealso{
Other Fitting Autoregressive Model Functions: 
\code{\link{FitVARLasso}()},
\code{\link{FitVAROLS}()},
\code{\link{LambdaSeq}()},
\code{\link{OrigScale}()},
\code{\link{PBootVARLasso}()},
\code{\link{PBootVAROLS}()},
\code{\link{RBootVARLasso}()},
\code{\link{RBootVAROLS}()},
\code{\link{SearchVARLasso}()},
\code{\link{StdMat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Fitting Autoregressive Model Functions}
\keyword{fit}
\keyword{fitAutoReg}
