% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitAutoReg-fit-var-mplus.R
\name{FitVARMplus}
\alias{FitVARMplus}
\title{Fit Vector Autoregressive Model using Mplus}
\usage{
FitVARMplus(data, p = 1, mplus_bin, iter = 5000L)
}
\arguments{
\item{data}{Numeric matrix.
The time series data with dimensions \code{t} by \code{k},
where \code{t} is the number of observations
and \code{k} is the number of variables.}

\item{p}{Positive integer.
Number of lags.}

\item{mplus_bin}{Character string.
Path to the Mplus binary.}

\item{iter}{Positive integer.
Maximum number of iterations
for each MCMC chain.}
}
\description{
This function estimates the parameters of a VAR model
using \code{Mplus}.
}
\examples{
\dontrun{
FitVARMplus(data = dat_p1, p = 1, mplus_bin = "mpdemo")
FitVARMplus(data = dat_p2, p = 2, mplus_bin = "mpdemo")
}

}
\seealso{
Other Fitting Autoregressive Model Functions: 
\code{\link{FitMLVARDynr}()},
\code{\link{FitMLVARMplus}()},
\code{\link{FitVARDynr}()},
\code{\link{FitVARLassoSearch}()},
\code{\link{FitVARLasso}()},
\code{\link{FitVAROLS}()},
\code{\link{LambdaSeq}()},
\code{\link{ModelVARP1Dynr}()},
\code{\link{ModelVARP2Dynr}()},
\code{\link{OrigScale}()},
\code{\link{PBootVARLasso}()},
\code{\link{PBootVAROLS}()},
\code{\link{RBootVARLasso}()},
\code{\link{RBootVAROLS}()},
\code{\link{SearchVARLasso}()},
\code{\link{StdMat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Fitting Autoregressive Model Functions}
\keyword{fit}
\keyword{fitAutoReg}
