% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LambdaSeq}
\alias{LambdaSeq}
\title{Function to generate the sequence of lambdas}
\usage{
LambdaSeq(YStd, XStd, n_lambdas)
}
\arguments{
\item{YStd}{Numeric matrix.
Matrix of standardized dependent variables (Y).}

\item{XStd}{Numeric matrix.
Matrix of standardized predictors (X).
\code{XStd} should not include a vector of ones in column one.}

\item{n_lambdas}{Integer.
Number of lambdas to generate.}
}
\value{
Returns a vector of lambdas.
}
\description{
Function to generate the sequence of lambdas
}
\examples{
YStd <- StdMat(dat_p2_yx$Y)
XStd <- StdMat(dat_p2_yx$X[, -1]) # remove the constant column
LambdaSeq(YStd = YStd, XStd = XStd, n_lambdas = 100)

}
\seealso{
Other Fitting Autoregressive Model Functions: 
\code{\link{FitMLVARDynr}()},
\code{\link{FitMLVARMplus}()},
\code{\link{FitVARDynr}()},
\code{\link{FitVARLassoSearch}()},
\code{\link{FitVARLasso}()},
\code{\link{FitVARMplus}()},
\code{\link{FitVAROLS}()},
\code{\link{ModelVARP1Dynr}()},
\code{\link{ModelVARP2Dynr}()},
\code{\link{OrigScale}()},
\code{\link{PBootVARLasso}()},
\code{\link{PBootVAROLS}()},
\code{\link{RBootVARLasso}()},
\code{\link{RBootVAROLS}()},
\code{\link{SearchVARLasso}()},
\code{\link{StdMat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Fitting Autoregressive Model Functions}
\keyword{fit}
\keyword{fitAutoReg}
