% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{OrigScale}
\alias{OrigScale}
\title{Return Standardized Estimates to the Original Scale}
\usage{
OrigScale(coef_std, Y, X)
}
\arguments{
\item{coef_std}{Numeric matrix.
Standardized estimates of the autoregression
and cross regression coefficients.}

\item{Y}{Numeric matrix.
Matrix of dependent variables (Y).}

\item{X}{Numeric matrix.
Matrix of predictors (X).}
}
\description{
Return Standardized Estimates to the Original Scale
}
\examples{
Y <- dat_demo_yx$Y
X <- dat_demo_yx$X[, -1]
Ystd <- StdMat(Y)
Xstd <- StdMat(X)
coef_std <- FitVAROLS(Y = Ystd, X = Xstd)
FitVAROLS(Y = Y, X = X)
OrigScale(coef_std = coef_std, Y = Y, X = X)

}
\seealso{
Other Fitting Autoregressive Model Functions: 
\code{\link{FitVARLassoSearch}()},
\code{\link{FitVARLasso}()},
\code{\link{FitVAROLS}()},
\code{\link{LambdaSeq}()},
\code{\link{PBootVARLasso}()},
\code{\link{PBootVAROLS}()},
\code{\link{SearchVARLasso}()},
\code{\link{StdMat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Fitting Autoregressive Model Functions}
\keyword{fitAutoReg}
\keyword{utils}
