% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitAutoReg-p-boot-ci.R
\name{PBootCI}
\alias{PBootCI}
\title{Parametric Bootstrap Confidence Intervals}
\usage{
PBootCI(x, alpha = 0.05)
}
\arguments{
\item{x}{Numeric matrix.
Output of \code{\link[=PBootVAROLS]{PBootVAROLS()}}.}

\item{alpha}{Numeric.
Significance level.}
}
\value{
A list with two elements, namely \code{ll} for the lower limit
and \code{ul} for the upper limit.
}
\description{
Parametric Bootstrap Confidence Intervals
}
\examples{
set.seed(42)
system.time(pb <- PBootVAROLS(data = dat_p2, p = 2, B = 10, burn_in = 20))
pb$est
PBootCI(pb)
system.time(pb <- PBootVARLasso(
  data = dat_p2, p = 2, B = 10, burn_in = 20,
  n_lambdas = 100, crit = "ebic", max_iter = 1000, tol = 1e-5
))
pb$est
PBootCI(pb)

}
\seealso{
Other Simulation of Autoregressive Data Functions: 
\code{\link{PBootSE}()},
\code{\link{SelectVARLasso}()},
\code{\link{SimVAR}()},
\code{\link{YX}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation of Autoregressive Data Functions}
\keyword{fitAutoReg}
\keyword{pb}
