% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitAutoReg-p-boot-se.R
\name{PBootSE}
\alias{PBootSE}
\title{Parametric Bootstrap Standard Errors}
\usage{
PBootSE(x)
}
\arguments{
\item{x}{Numeric matrix.
Output of \code{\link[=PBootVAROLS]{PBootVAROLS()}}.}
}
\value{
A matrix of standard error.
}
\description{
Parametric Bootstrap Standard Errors
}
\examples{
set.seed(42)
system.time(pb <- PBootVAROLS(data = dat_demo, p = 2, B = 10, burn_in = 20))
pb$est
PBootSE(pb)
system.time(pb <- PBootVARLasso(
  data = dat_demo, p = 2, B = 10, burn_in = 20,
  n_lambdas = 100, crit = "ebic", max_iter = 1000, tol = 1e-5
))
pb$est
PBootSE(pb)

}
\seealso{
Other Simulation of Autoregressive Data Functions: 
\code{\link{PBootCI}()},
\code{\link{SelectVARLasso}()},
\code{\link{SimVAR}()},
\code{\link{YX}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation of Autoregressive Data Functions}
\keyword{fitAutoReg}
\keyword{pb}
