% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PBootVARExoOLS}
\alias{PBootVARExoOLS}
\title{Parametric Bootstrap for the Vector Autoregressive Model
with Exogenous Variables
Using Ordinary Least Squares}
\usage{
PBootVARExoOLS(data, exo_mat, p, B, burn_in)
}
\arguments{
\item{data}{Numeric matrix.
The time series data with dimensions \code{t} by \code{k},
where \code{t} is the number of observations
and \code{k} is the number of variables.}

\item{exo_mat}{Numeric matrix.
Matrix of exogenous variables with dimensions \code{t + burn_in} by \code{m}.
If the number of rows is equal to \code{t}, set \code{burn_in = 0}.}

\item{p}{Integer.
The order of the VAR model (number of lags).}

\item{B}{Integer.
Number of bootstrap samples to generate.}

\item{burn_in}{Integer.
Number of burn-in observations to exclude before returning the results
in the simulation step.}
}
\value{
List with the following elements:
\itemize{
\item \strong{est}: Numeric matrix.
Original OLS estimate of the coefficient matrix.
\item \strong{boot}: Numeric matrix.
Matrix of vectorized bootstrap estimates of the coefficient matrix.
}
}
\description{
Parametric Bootstrap for the Vector Autoregressive Model
with Exogenous Variables
Using Ordinary Least Squares
}
\examples{
data <- dat_p2_exo$data
exo_mat <- dat_p2_exo$exo_mat
pb <- PBootVARExoOLS(
  data = data,
  exo_mat = exo_mat,
  p = 2,
  B = 5,
  burn_in = 0
)
str(pb)

}
\seealso{
Other Fitting Autoregressive Model Functions: 
\code{\link{FitMLVARDynr}()},
\code{\link{FitMLVARMplus}()},
\code{\link{FitVARDynr}()},
\code{\link{FitVARLassoSearch}()},
\code{\link{FitVARLasso}()},
\code{\link{FitVARMplus}()},
\code{\link{FitVAROLS}()},
\code{\link{LambdaSeq}()},
\code{\link{ModelVARP1Dynr}()},
\code{\link{ModelVARP2Dynr}()},
\code{\link{OrigScale}()},
\code{\link{PBootVARExoLasso}()},
\code{\link{PBootVARLasso}()},
\code{\link{PBootVAROLS}()},
\code{\link{RBootVARExoLasso}()},
\code{\link{RBootVARExoOLS}()},
\code{\link{RBootVARLasso}()},
\code{\link{RBootVAROLS}()},
\code{\link{SearchVARLasso}()},
\code{\link{StdMat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Fitting Autoregressive Model Functions}
\keyword{fitAutoReg}
\keyword{pb}
