% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PBootVARLasso}
\alias{PBootVARLasso}
\title{Parametric Bootstrap for the Vector Autoregressive Model
Using Lasso Regularization}
\usage{
PBootVARLasso(data, p, B, burn_in, n_lambdas, crit, max_iter, tol)
}
\arguments{
\item{data}{Numeric matrix.
The time series data with dimensions \code{t} by \code{k},
where \code{t} is the number of observations
and \code{k} is the number of variables.}

\item{p}{Integer.
The order of the VAR model (number of lags).}

\item{B}{Integer.
Number of bootstrap samples to generate.}

\item{burn_in}{Integer.
Number of burn-in observations to exclude before returning the results
in the simulation step.}

\item{n_lambdas}{Integer.
Number of lambdas to generate.}

\item{crit}{Character string.
Information criteria to use.
Valid values include \code{"aic"}, \code{"bic"}, and \code{"ebic"}.}

\item{max_iter}{Integer.
The maximum number of iterations for the coordinate descent algorithm
(e.g., \code{max_iter = 10000}).}

\item{tol}{Numeric.
Convergence tolerance. The algorithm stops when the change in coefficients
between iterations is below this tolerance
(e.g., \code{tol = 1e-5}).}
}
\value{
List containing the estimates (\code{est})
and bootstrap estimates (\code{boot}).
}
\description{
Parametric Bootstrap for the Vector Autoregressive Model
Using Lasso Regularization
}
\examples{
pb <- PBootVARLasso(data = dat_p2, p = 2, B = 10, burn_in = 20,
  n_lambdas = 100, crit = "ebic", max_iter = 1000, tol = 1e-5)
str(pb)

}
\seealso{
Other Fitting Autoregressive Model Functions: 
\code{\link{FitMLVARDynr}()},
\code{\link{FitMLVARMplus}()},
\code{\link{FitVARDynr}()},
\code{\link{FitVARLassoSearch}()},
\code{\link{FitVARLasso}()},
\code{\link{FitVARMplus}()},
\code{\link{FitVAROLS}()},
\code{\link{LambdaSeq}()},
\code{\link{ModelVARP1Dynr}()},
\code{\link{ModelVARP2Dynr}()},
\code{\link{OrigScale}()},
\code{\link{PBootVAROLS}()},
\code{\link{SearchVARLasso}()},
\code{\link{StdMat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Fitting Autoregressive Model Functions}
\keyword{fitAutoReg}
\keyword{pb}
