% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PBootVAROLS}
\alias{PBootVAROLS}
\title{Parametric Bootstrap for the Vector Autoregressive Model
Using Ordinary Least Squares}
\usage{
PBootVAROLS(data, p, B, burn_in)
}
\arguments{
\item{data}{Numeric matrix.
The time series data with dimensions \code{t} by \code{k},
where \code{t} is the number of observations
and \code{k} is the number of variables.}

\item{p}{Integer.
The order of the VAR model (number of lags).}

\item{B}{Integer.
Number of bootstrap samples to generate.}

\item{burn_in}{Integer.
Number of burn-in observations to exclude before returning the results
in the simulation step.}
}
\value{
List with the following elements:
\itemize{
\item \strong{est}: Numeric matrix.
Original OLS estimate of the coefficient matrix.
\item \strong{boot}: Numeric matrix.
Matrix of vectorized bootstrap estimates of the coefficient matrix.
}
}
\description{
Parametric Bootstrap for the Vector Autoregressive Model
Using Ordinary Least Squares
}
\examples{
PBootVAROLS(data = dat_p2, p = 2, B = 5, burn_in = 20)

}
\seealso{
Other Fitting Autoregressive Model Functions: 
\code{\link{FitVARLassoSearch}()},
\code{\link{FitVARLasso}()},
\code{\link{FitVAROLS}()},
\code{\link{LambdaSeq}()},
\code{\link{OrigScale}()},
\code{\link{PBootVARLasso}()},
\code{\link{RBootVARExoLasso}()},
\code{\link{RBootVARExoOLS}()},
\code{\link{RBootVARLasso}()},
\code{\link{RBootVAROLS}()},
\code{\link{SearchVARLasso}()},
\code{\link{StdMat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Fitting Autoregressive Model Functions}
\keyword{fitAutoReg}
\keyword{pb}
