% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitAutoReg-methods.R
\name{plot.dynr_model}
\alias{plot.dynr_model}
\title{Plot Method for an Object of Class \code{dynr_model}}
\usage{
\method{plot}{dynr_model}(x, ...)
}
\arguments{
\item{x}{Object of class \code{dynr_model}.}

\item{...}{additional arguments.}
}
\description{
Plot Method for an Object of Class \code{dynr_model}
}
\examples{
model <- ModelVARP2Dynr(data = dat_demo)
plot(model)

}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
