% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitAutoReg-methods.R
\name{print.dynr_model}
\alias{print.dynr_model}
\title{Print Method for an Object of Class \code{dynr_model}}
\usage{
\method{print}{dynr_model}(x, ...)
}
\arguments{
\item{x}{Object of class \code{dynr_model}.}

\item{...}{additional arguments.}
}
\description{
Print Method for an Object of Class \code{dynr_model}
}
\examples{
model <- ModelVARP2Dynr(data = dat_demo)
print(model)

}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
