% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RBootVARExoOLS}
\alias{RBootVARExoOLS}
\title{Residual Bootstrap for the Vector Autoregressive Model
with Exogenous Variables
Using Ordinary Least Squares}
\usage{
RBootVARExoOLS(data, exo_mat, p, B)
}
\arguments{
\item{data}{Numeric matrix.
The time series data with dimensions \code{t} by \code{k},
where \code{t} is the number of observations
and \code{k} is the number of variables.}

\item{exo_mat}{Numeric matrix.
Matrix of exogenous variables with dimensions \code{t} by \code{m}.}

\item{p}{Integer.
The order of the VAR model (number of lags).}

\item{B}{Integer.
Number of bootstrap samples to generate.}
}
\value{
List with the following elements:
\itemize{
\item \strong{est}: Numeric matrix.
Original OLS estimate of the coefficient matrix.
\item \strong{boot}: Numeric matrix.
Matrix of vectorized bootstrap estimates of the coefficient matrix.
\item \strong{X}: Numeric matrix.
Original \code{X}
\item \strong{Y}: List of numeric matrices.
Bootstrapped \code{Y}
}
}
\description{
Residual Bootstrap for the Vector Autoregressive Model
with Exogenous Variables
Using Ordinary Least Squares
}
\examples{
data <- dat_p2_exo$data
exo_mat <- dat_p2_exo$exo_mat
RBootVARExoOLS(data = data, exo_mat = exo_mat, p = 2, B = 5)

}
\seealso{
Other Fitting Autoregressive Model Functions: 
\code{\link{FitVARLassoSearch}()},
\code{\link{FitVARLasso}()},
\code{\link{FitVAROLS}()},
\code{\link{LambdaSeq}()},
\code{\link{OrigScale}()},
\code{\link{PBootVARLasso}()},
\code{\link{PBootVAROLS}()},
\code{\link{RBootVARExoLasso}()},
\code{\link{RBootVARLasso}()},
\code{\link{RBootVAROLS}()},
\code{\link{SearchVARLasso}()},
\code{\link{StdMat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Fitting Autoregressive Model Functions}
\keyword{fitAutoReg}
\keyword{rb}
