% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RBootVAROLS}
\alias{RBootVAROLS}
\title{Residual Bootstrap for the Vector Autoregressive Model
Using Ordinary Least Squares}
\usage{
RBootVAROLS(data, p, B)
}
\arguments{
\item{data}{Numeric matrix.
The time series data with dimensions \code{t} by \code{k},
where \code{t} is the number of observations
and \code{k} is the number of variables.}

\item{p}{Integer.
The order of the VAR model (number of lags).}

\item{B}{Integer.
Number of bootstrap samples to generate.}
}
\value{
List with the following elements:
\itemize{
\item \strong{est}: Numeric matrix.
Original OLS estimate of the coefficient matrix.
\item \strong{boot}: Numeric matrix.
Matrix of vectorized bootstrap estimates of the coefficient matrix.
\item \strong{X}: Numeric matrix.
Original \code{X}
\item \strong{Y}: List of numeric matrices.
Bootstrapped \code{Y}
}
}
\description{
Residual Bootstrap for the Vector Autoregressive Model
Using Ordinary Least Squares
}
\examples{
rb <- RBootVAROLS(data = dat_p2, p = 2, B = 5)
str(rb)

}
\seealso{
Other Fitting Autoregressive Model Functions: 
\code{\link{FitMLVARDynr}()},
\code{\link{FitMLVARMplus}()},
\code{\link{FitVARDynr}()},
\code{\link{FitVARLassoSearch}()},
\code{\link{FitVARLasso}()},
\code{\link{FitVARMplus}()},
\code{\link{FitVAROLS}()},
\code{\link{LambdaSeq}()},
\code{\link{ModelVARP1Dynr}()},
\code{\link{ModelVARP2Dynr}()},
\code{\link{OrigScale}()},
\code{\link{PBootVARExoLasso}()},
\code{\link{PBootVARExoOLS}()},
\code{\link{PBootVARLasso}()},
\code{\link{PBootVAROLS}()},
\code{\link{RBootVARExoLasso}()},
\code{\link{RBootVARExoOLS}()},
\code{\link{RBootVARLasso}()},
\code{\link{SearchVARLasso}()},
\code{\link{StdMat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Fitting Autoregressive Model Functions}
\keyword{fitAutoReg}
\keyword{rb}
