% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FitVARLassoSearch}
\alias{FitVARLassoSearch}
\title{Fit Vector Autoregressive (VAR) Model Parameters using Lasso Regularization
with Lambda Search}
\usage{
FitVARLassoSearch(Ystd, Xstd, lambdas, crit, max_iter, tol)
}
\arguments{
\item{Ystd}{Numeric matrix.
Matrix of standardized dependent variables (Y).}

\item{Xstd}{Numeric matrix.
Matrix of standardized predictors (X).
\code{Xstd} should not include a vector of ones in column one.}

\item{lambdas}{Numeric vector.
Vector of lambda hyperparameters for Lasso regularization.}

\item{crit}{Character string.
Information criteria to use.
Valid values include \code{"aic"}, \code{"bic"}, and \code{"ebic"}.}

\item{max_iter}{Integer.
The maximum number of iterations for the coordinate descent algorithm
(e.g., \code{max_iter = 10000}).}

\item{tol}{Numeric.
Convergence tolerance. The algorithm stops when the change in coefficients
between iterations is below this tolerance
(e.g., \code{tol = 1e-5}).}
}
\value{
Matrix of estimated autoregressive
and cross-regression coefficients.
}
\description{
Fit Vector Autoregressive (VAR) Model Parameters using Lasso Regularization
with Lambda Search
}
\examples{
Ystd <- StdMat(dat_p2_yx$Y)
Xstd <- StdMat(dat_p2_yx$X[, -1])
lambdas <- LambdaSeq(Y = Ystd, X = Xstd, n_lambdas = 100)
FitVARLassoSearch(Ystd = Ystd, Xstd = Xstd, lambdas = lambdas,
  crit = "ebic", max_iter = 1000, tol = 1e-5)

}
\seealso{
Other Fitting Autoregressive Model Functions: 
\code{\link{FitMLVARDynr}()},
\code{\link{FitMLVARMplus}()},
\code{\link{FitVARDynr}()},
\code{\link{FitVARLasso}()},
\code{\link{FitVARMplus}()},
\code{\link{FitVAROLS}()},
\code{\link{LambdaSeq}()},
\code{\link{ModelVARP1Dynr}()},
\code{\link{ModelVARP2Dynr}()},
\code{\link{OrigScale}()},
\code{\link{PBootVARLasso}()},
\code{\link{PBootVAROLS}()},
\code{\link{RBootVARLasso}()},
\code{\link{RBootVAROLS}()},
\code{\link{SearchVARLasso}()},
\code{\link{StdMat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Fitting Autoregressive Model Functions}
\keyword{fit}
\keyword{fitAutoReg}
