% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LambdaSeq}
\alias{LambdaSeq}
\title{Function to generate the sequence of lambdas}
\usage{
LambdaSeq(Y, X, n_lambdas)
}
\arguments{
\item{Y}{Numeric matrix.
Matrix of dependent variables (Y).}

\item{X}{Numeric matrix.
Matrix of predictors (X).}

\item{n_lambdas}{Integer.
Number of lambdas to generate.}
}
\value{
Returns a vector of lambdas.
}
\description{
Function to generate the sequence of lambdas
}
\examples{
Ystd <- StdMat(dat_demo_yx$Y)
Xstd <- StdMat(dat_demo_yx$X[, -1])
LambdaSeq(Y = Ystd, X = Xstd, n_lambdas = 100)

}
\seealso{
Other Fitting Autoregressive Model Functions: 
\code{\link{FitMLVARDynr}()},
\code{\link{FitVARDynr}()},
\code{\link{FitVARLassoSearch}()},
\code{\link{FitVARLasso}()},
\code{\link{FitVAROLS}()},
\code{\link{ModelVARP1Dynr}()},
\code{\link{ModelVARP2Dynr}()},
\code{\link{OrigScale}()},
\code{\link{PBootVARLasso}()},
\code{\link{PBootVAROLS}()},
\code{\link{SearchVARLasso}()},
\code{\link{StdMat}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Fitting Autoregressive Model Functions}
\keyword{fit}
\keyword{fitAutoReg}
