% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitAutoReg-select-var-lasso.R
\name{SelectVARLasso}
\alias{SelectVARLasso}
\title{Select the Lasso Estimates from the Grid Search}
\usage{
SelectVARLasso(search, crit = "ebic")
}
\arguments{
\item{search}{Object.
Output of the \code{\link[=SearchVARLasso]{SearchVARLasso()}} function.}

\item{crit}{Character string.
Information criteria to use.
Valid values include \code{"aic"}, \code{"bic"}, and \code{"ebic"}.}
}
\value{
Returns the Lasso estimates
of autoregression and cross regression coefficients.
}
\description{
Select the Lasso Estimates from the Grid Search
}
\examples{
YStd <- StdMat(dat_p2_yx$Y)
XStd <- StdMat(dat_p2_yx$X[, -1])
lambdas <- 10^seq(-5, 5, length.out = 100)
search <- SearchVARLasso(
  YStd = YStd, XStd = XStd, lambdas = lambdas,
  max_iter = 10000, tol = 1e-5
)
SelectVARLasso(search, crit = "ebic")

}
\seealso{
Other Simulation of Autoregressive Data Functions: 
\code{\link{BootCI}()},
\code{\link{BootSE}()},
\code{\link{SimVAR}()},
\code{\link{YX}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation of Autoregressive Data Functions}
\keyword{fit}
\keyword{fitAutoReg}
