% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{StdMat}
\alias{StdMat}
\title{Standardize Matrix}
\usage{
StdMat(X)
}
\arguments{
\item{X}{Numeric matrix.
The matrix to be standardized.}
}
\value{
Numeric matrix with standardized values.
}
\description{
This function standardizes the given matrix by centering the columns
and scaling them to have unit variance.
}
\examples{
std <- StdMat(dat_demo)
colMeans(std)
var(std)

}
\seealso{
Other Fitting Autoregressive Model Functions: 
\code{\link{FitMLVARDynr}()},
\code{\link{FitVARDynr}()},
\code{\link{FitVARLassoSearch}()},
\code{\link{FitVARLasso}()},
\code{\link{FitVAROLS}()},
\code{\link{LambdaSeq}()},
\code{\link{ModelVARP1Dynr}()},
\code{\link{ModelVARP2Dynr}()},
\code{\link{OrigScale}()},
\code{\link{PBootVARLasso}()},
\code{\link{PBootVAROLS}()},
\code{\link{SearchVARLasso}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Fitting Autoregressive Model Functions}
\keyword{fitAutoReg}
\keyword{utils}
